Refactor task metrics
